requirements.txt
Fabric==1.12.0
python-slugify==1.2.1
requests==2.13.0